management users handled close archivados usuarios #67
