+++ src/app/dashboard/management/users/components/DeleteEntityDialog.tsx
@@ -1,12 +1,15 @@
 'use client'
 
-import { deleteUser } from '@/actions/users'
+import { getAdminRole } from '@/actions/roles'
+import { adminCount, deleteUser } from '@/actions/users'
 import { Button } from '@/components/Button'
 import { Dialog, DialogContent, DialogTitle } from '@/components/Dialog'
 import {
-    openDeleteUserAtom,
-    updateUsersAtom,
-    userToEditAtom,
+    openDeleteAtom,
+    updateAtom,
+    entityToEditAtom,
+    openPreventArchiveAdminAtom,
+    adminRoleAtom,
 } from '@/global/management-users'
 import { DialogDescription } from '@radix-ui/react-dialog'
 import { useAtom, useAtomValue, useSetAtom } from 'jotai'
@@ -14,13 +17,15 @@ import { Ban, Trash2 } from 'lucide-react'
 import { useState, useTransition } from 'react'
 
 export function DeleteEntityDialog() {
-    const [open, setOpen] = useAtom(openDeleteUserAtom)
+    const [open, setOpen] = useAtom(openDeleteAtom)
     const [inTransition, startTransition] = useTransition()
-    const entity = useAtomValue(userToEditAtom)
+    const entity = useAtomValue(entityToEditAtom)
     const [message, setMessage] = useState('')
-    const updateUsersTable = useSetAtom(updateUsersAtom)
+    const updateUsersTable = useSetAtom(updateAtom)
+    const setOpenPreventArchiveAdmin = useSetAtom(openPreventArchiveAdminAtom)
+    const adminRole = useAtomValue(adminRoleAtom)
 
-    if (!entity) return null
+    if (!entity || !adminRole) return null
 
     return (
         <Dialog open={open} onOpenChange={setOpen}>
@@ -36,6 +41,14 @@ export function DeleteEntityDialog() {
                 <form
                     action={data => {
                         startTransition(async () => {
+                            if (entity.role_id === adminRole.id) {
+                                const ac = await adminCount()
+                                if (ac < 2) {
+                                    setOpenPreventArchiveAdmin(true)
+                                    setOpen(false)
+                                    return
+                                }
+                            }
                             const { error } = await deleteUser(data)
                             if (error) {
                                 setMessage(
