+++ src/app/dashboard/management/users/components/UnarchiveOrDeleteDialog.tsx
@@ -3,35 +3,35 @@
 import { unarchiveUser } from '@/actions/users'
 import { Button } from '@/components/Button'
 import { Dialog, DialogContent, DialogTitle } from '@/components/Dialog'
+import { MessageError } from '@/components/Error'
 import {
-    openUnarchiveUserAtom,
-    updateAtom,
     entityToEditAtom,
+    updateAtom,
+    openUnarchiveOrDeleteAtom,
+    openDeleteAtom,
 } from '@/global/management-users'
 import { DialogDescription } from '@radix-ui/react-dialog'
 import { useAtom, useAtomValue, useSetAtom } from 'jotai'
-import { ArchiveRestore, Ban } from 'lucide-react'
+import { ArchiveRestoreIcon, BanIcon, TrashIcon } from 'lucide-react'
 import { useState, useTransition } from 'react'
 
-export function UnarchiveEntityDialog() {
-    const [open, setOpen] = useAtom(openUnarchiveUserAtom)
+export function UnarchiveOrDeleteDialog() {
+    const [open, setOpen] = useAtom(openUnarchiveOrDeleteAtom)
     const [inTransition, startTransition] = useTransition()
     const entity = useAtomValue(entityToEditAtom)
     const [message, setMessage] = useState('')
     const updateUsersTable = useSetAtom(updateAtom)
+    const setOpenDelete = useSetAtom(openDeleteAtom)
 
     if (!entity) return null
 
     return (
         <Dialog open={open} onOpenChange={setOpen}>
             <DialogContent>
-                <DialogTitle>
-                    <span className='text-3xl'>
-                        Desarchivar @{entity.username}
-                    </span>
-                </DialogTitle>
+                <DialogTitle>El usuario se encuentra archivado</DialogTitle>
                 <DialogDescription>
-                    ¿Está seguro de desarchivar a {entity.name}?
+                    Que quieres hacer con{' '}
+                    <span className='font-bold'>{entity.name}</span>?
                 </DialogDescription>
                 <form
                     action={data => {
@@ -51,31 +51,41 @@ export function UnarchiveEntityDialog() {
                     }}
                     className='flex w-full max-w-md flex-col justify-center gap-6'
                 >
-                    {message && (
-                        <span className='animate-slide-in mt-1 block rounded-lg bg-red-100 px-3 py-1 text-sm text-red-600 shadow-md'>
-                            {message}
-                        </span>
-                    )}
+                    {message && <MessageError>{message}</MessageError>}
                     <input type='hidden' value={entity.id} name='id' />
                     <div className='flex flex-row gap-2 *:flex-1'>
                         <Button
-                            variant={'secondary'}
+                            type='button'
+                            variant='secondary'
                             disabled={inTransition}
                             onClick={e => {
                                 e.preventDefault()
                                 setOpen(false)
                             }}
                         >
-                            <Ban className='mr-2 h-5 w-5' />
+                            <BanIcon className='mr-2 h-5 w-5' />
                             Cancelar
                         </Button>
                         <Button
                             type='submit'
-                            variant={'default'}
+                            variant='default'
                             disabled={inTransition}
                         >
-                            <ArchiveRestore className='mr-2 h-5 w-5' />
+                            <ArchiveRestoreIcon className='mr-2 h-5 w-5' />
                             Desarchivar
+                        </Button>
+                        <Button
+                            type='button'
+                            variant='destructive'
+                            disabled={inTransition}
+                            onClick={e => {
+                                e.preventDefault()
+                                setOpen(false)
+                                setOpenDelete(true)
+                            }}
+                        >
+                            <TrashIcon className='mr-2 h-5 w-5' />
+                            Eliminar
                         </Button>
                     </div>
                 </form>
