+++ src/app/dashboard/management/users/page.tsx
@@ -2,7 +2,7 @@ import { DashboardHeader } from '../../components/DashboardHeader'
 import { Filters } from './components/SearchInput'
 import { UsersTable } from './components/UsersTable'
 import { CreateButton } from './components/CreateButton'
-import { CreateUserDialog } from './components/CreateUserDialog'
+import { CreateUserDialog } from './components/CreateDialog'
 
 export default async function UserManagementPage() {
     return (
